requirements.txt
fastapi==0.110.0
alembic==1.13.1
cryptography==40.0.2
python-jose==3.3.0
pydantic==2.6.4
SQLAlchemy==2.0.29
psycopg2-binary==2.9.6
python-multipart==0.0.9
passlib==1.7.4
gunicorn==20.1.0
uvicorn==0.22.0
